Backend/api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
pillow==10.4.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
deprecation==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
exceptiongroup==1.2.1
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.0
postgrest==0.16.8
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.5
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
typing_extensions==4.12.1
websockets==12.0
